+++ memory.h
@@ -2,5 +2,5 @@
 
 #include "types.h"
 
-void* memset(void* dst, uint8_t c, uint64_t size);
-void* memmove(void* dst, const void* src, uint64_t size);
+void* memset(void* dst, uint8_t c, uint32_t size);
+void* memmove(void* dst, const void* src, uint32_t size);
